+++ server_root/public_html/cron/update-tutor-reports.php
@@ -11,14 +11,10 @@ $appointments = AppointmentFetcher::retrieveCompletedWithoutReports($db);
 var_dump($appointments);
 
 foreach ($appointments as $appointment) {
+	$appointment = $appointments[0];
 
-	// create report with
-	Report::
-//	$appointmentHasStudents = AppointmentHasStudentFetcher::retrieveJoinReport($db, $appointment[AppointmentFetcher::DB_COLUMN_ID]);
-//	var_dump($appointmentHasStudents);
-
-//	foreach()
-	// create report -- tutor can edit on his page
-//	Report::insert($db, $appointment[AppointmentFetcher::DB_COLUMN_ID], $appointment[AppointmentFetcher::DB])
-	// mail tutor
+	$reportId = ReportFetcher::insert($db, $appointment[AppointmentHasStudentFetcher::DB_COLUMN_STUDENT_ID],
+		$appointment[AppointmentHasStudentFetcher::DB_TABLE . "_" . AppointmentHasStudentFetcher::DB_COLUMN_ID],
+		$appointment[AppointmentHasStudentFetcher::DB_COLUMN_INSTRUCTOR_ID]);
+	Mailer::sendTutorNewReport($db, $reportId, $appointment[AppointmentFetcher::DB_COLUMN_TUTOR_USER_ID], $appointment[AppointmentFetcher::DB_COLUMN_COURSE_ID]);
 }