+++ server_root/app/plugins/PHPMailer/changelog.md
@@ -9,6 +9,15 @@
 * More consistent handling of error states and debug output
 * Use property defaults, remove constructors
 * Remove unreachable code
+* Use older regex validation pattern for troublesome PCRE library versions
+* Improve PCRE detection in older PHP versions
+* Handle debug output consistently, and always in UTF-8
+* Allow user-defined debug output method via a callable
+* msgHTML now converts data URIs to embedded images
+* SMTP::getLastReply() will now always be populated
+* The autoloader is no longer autoloaded by the PHPMailer class
+* Improved example code in README
+* Ensure long filenames in Content-Disposition are encoded correctly
 
 ## Version 5.2.8 (May 14th 2014)
 * Increase timeout to match RFC2821 section 4.5.3.2 and thus not fail greetdelays, fixes #104